refactor: `Repository` 인터페이스를 도메인으로 이동, 구현체와 분리
