+++ src/main/java/roomescape/domain/TimeSlotRepository.java
@@ -1,9 +1,8 @@
-package roomescape.repository;
+package roomescape.domain;
 
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.TimeSlot;
 
 public interface TimeSlotRepository {
